requirements.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.11.0
boto3==1.28.30
botocore==1.31.30
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.0
cron-descriptor==1.4.0
cryptography==41.0.3
decorator==5.1.1
dill==0.3.7
Django==4.0.4
django-appconf==1.0.5
django-bootstrap-form==3.4
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-json-widget==1.1.1
django-model-utils==4.3.1
django-polymorphic==3.1.0
django-redis==5.3.0
django-redis-cache==3.0.1
django-redis-sessions==0.6.2
django-select2==8.1.2
django-statsd-mozilla==0.4.0
django-storages==1.13.2
django-swagger-ui==0.1.11
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
exceptiongroup==1.1.3
executing==1.2.0
factory-boy==3.3.0
Faker==19.3.0
flake8==4.0.1
flower==2.0.1
future==0.18.3
google-auth==2.22.0
google-oauth==1.0.1
greenlet==2.0.2
httplib2==0.22.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.14.0
isort==5.12.0
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
josepy==1.13.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
lazy-object-proxy==1.9.0
magicmock==0.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mozilla-django-oidc==3.0.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
ordered-set==4.1.0
packaging==23.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pyOpenSSL==23.2.0
pyparsing==3.1.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-flake8==1.1.1
pytest-isort==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-memcached==1.59
pytz==2023.3
PyYAML==6.0.1
redis==3.5.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.2
sentry-sdk==1.29.2
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
statsd==4.0.0
svgwrite==1.4.3
tblib==2.0.0
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
vine==5.0.0
watchtower==3.0.0
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.16.2
